requirements.txt
django-ninja==0.22.2
django-ninja-extra==0.19.1
pydantic==1.9.1
django-cors-headers==4.2.0
requests==2.31.0
psycopg==3.1.10
psycopg2-binary==2.9.5
celery==5.3.4
redis==4.6.0
python-decouple==3.8
django-storages==1.14.2
firebase-admin==6.2.0